twitter-aggregator/requirements.txt
tweepy==4.14.0
requests==2.31.0
nltk==3.8.1
spacy==3.7.2
flask-cors
python-dotenv==1.0.0
firebase-admin==6.2.0 